Added a function and named cleaned_players and added comments to clarify certain pieces of logic.
cleaned_players takes in a list of dictionaries and returns a list of dictionaries with new values for each key of each dictionary based off of some prerequisite condition. This function will be used on the PLAYERS constant imported from constants.py.

Some comments were added to clarify logic within in the method and script in general.