Added some needed functionality
Created three functions one for the first menu prompt, another for the second menu prompt, and another to output the stats of the choosen team.

Added another function to calculate average heigth, and to test if a users input was valid.

I added additional comments to the script to explain logic.